Legg til mulighet for forskjellige endrede perioder. Splitt ut vedtaksperiodene i egne filer.
