+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/dokument/domene/maler/Avslag.kt
@@ -1,5 +1,7 @@
 package no.nav.familie.ba.sak.kjerne.dokument.domene.maler
 
+import no.nav.familie.ba.sak.kjerne.dokument.domene.maler.brevperioder.BrevPeriode
+
 data class Avslag(
     override val mal: Brevmal = Brevmal.VEDTAK_AVSLAG,
     override val data: AvslagData
