+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/dokument/domene/maler/Opphørt.kt
@@ -1,5 +1,7 @@
 package no.nav.familie.ba.sak.kjerne.dokument.domene.maler
 
+import no.nav.familie.ba.sak.kjerne.dokument.domene.maler.brevperioder.BrevPeriode
+
 data class Opphørt(
     override val mal: Brevmal = Brevmal.VEDTAK_OPPHØRT,
     override val data: OpphørtData
