+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/dokument/domene/maler/Vedtaksbrev.kt
@@ -1,5 +1,7 @@
 package no.nav.familie.ba.sak.kjerne.dokument.domene.maler
 
+import no.nav.familie.ba.sak.kjerne.dokument.domene.maler.brevperioder.BrevPeriode
+
 interface Vedtaksbrev : Brev {
 
     override val mal: Brevmal
@@ -17,7 +19,12 @@ enum class BrevPeriodeType(val apiNavn: String) {
     AVSLAG("avslag"),
     AVSLAG_UTEN_PERIODE("avslagUtenPeriode"),
     FORTSATT_INNVILGET("fortsattInnvilget"),
+}
+
+enum class EndretUtbetalingBrevPeriodeType(val apiNavn: String) {
     ENDRET_UTBETALINGSPERIODE("endretUtbetalingsperiode"),
+    ENDRET_UTBETALINGSPERIODE_DELVIS_UTBETALING("endretUtbetalingsperiodeDelvisUtbetaling"),
+    ENDRET_UTBETALINGSPERIODE_INGEN_UTBETALING("endretUtbetalingsperiodeIngenUtbetaling"),
 }
 
 data class VedtakFellesfelter(
